Add defaultRevision logic to tagWatcher (#57467)
* Add defaultRevision logic to tagWatcher

The defaultRevision should be treated equivalently to the default tag,
so update the tag watcher to look for the default revision as well.


* Only check the default revision if the rev label is unset

* Handle conflicts if default revision and default tag are different


---------

+++ tests/integration/helm/install_test.go
@@ -24,11 +24,15 @@ import (
 	"strings"
 	"testing"
 
+	klabels "k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/selection"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/yaml"
 
+	"istio.io/api/label"
 	"istio.io/istio/pkg/test/framework"
 	kubecluster "istio.io/istio/pkg/test/framework/components/cluster/kube"
+	"istio.io/istio/pkg/test/framework/components/crd"
 	"istio.io/istio/pkg/test/framework/components/namespace"
 	"istio.io/istio/pkg/test/helm"
 	"istio.io/istio/tests/util/sanitycheck"
@@ -42,6 +46,43 @@ func TestDefaultInstall(t *testing.T) {
 	framework.
 		NewTest(t).
 		Run(setupInstallation(values, false, DefaultNamespaceConfig, ""))
+}
+
+func TestRevisionedInstall(t *testing.T) {
+	values := map[string]interface{}{
+		"global":          map[string]interface{}{},
+		"defaultRevision": "testrev",
+		"revision":        "testrev",
+	}
+	revision := "testrev"
+	framework.
+		NewTest(t).
+		Run(baseSetup(values, false, DefaultNamespaceConfig, func(t framework.TestContext) {
+			// Install gateway API CRDs
+			crd.DeployGatewayAPI(t)
+			// Verify we can create a Gateway successfully
+			sampleGateway := `
+apiVersion: gateway.networking.k8s.io/v1
+kind: Gateway
+metadata:
+  name: sample
+  namespace: default
+spec:
+  gatewayClassName: istio
+  listeners:
+  - name: http
+    protocol: HTTP
+    port: 80
+`
+			t.ConfigIstio().Eval("default", nil, fmt.Sprint(sampleGateway)).ApplyOrFail(t)
+			selector := klabels.NewSelector()
+			req, _ := klabels.NewRequirement(label.IoK8sNetworkingGatewayGatewayName.Name, selection.Equals, []string{"sample"})
+			selector.Add(*req)
+			VerifyPodReady(t, t.Clusters().Default(), "default", selector.String())
+			if !t.Settings().NoCleanup {
+				t.ConfigIstio().Eval("default", nil, fmt.Sprint(sampleGateway)).DeleteOrFail(t)
+			}
+		}, revision))
 }
 
 // TestAmbientInstall tests Istio ambient profile installation using Helm
@@ -240,9 +281,12 @@ func baseSetup(values map[string]interface{}, isAmbient bool, config NamespaceCo
 		VerifyInstallation(t, cs, config, true, isAmbient, revision)
 		verifyValidation(t, revision)
 
-		check(t)
 		t.Cleanup(func() {
-			DeleteIstio(t, h, cs, config, isAmbient)
+			if !t.Settings().NoCleanup {
+				DeleteIstio(t, h, cs, config, isAmbient)
+			}
 		})
+
+		check(t)
 	}
 }
